+++ src/pages/Home/CommonWalkingMoneyRoad.tsx
