+++ src/pages/Home/ParentPendingMoneyRoad.tsx
@@ -1,0 +1,5 @@
+function ParentPendingMoneyRoad() {
+  return <>부모 - 챌린지 요청</>;
+}
+
+export default ParentPendingMoneyRoad;
