--- src/pages/Home/kid/KidPendingMoneyRoad.tsx
@@ -1,5 +1,0 @@
-function PendingKid() {
-  return <>자녀 - 챌린지 요청</>;
-}
-
-export default PendingKid;
