--- src/pages/Home/parent/ParentPendingMoneyRoad.tsx
@@ -1,5 +1,0 @@
-function PendingParent() {
-  return <>부모 - 챌린지 요청</>;
-}
-
-export default PendingParent;
